+++ packages/sdk/src/types/Timestamp.ts
@@ -10,9 +10,6 @@ import {
  * @deprecated
  */
 export const JSON_DISCRIMINATOR = TypedJsonDiscriminator.Timestamp;
-/**
- * @deprecated
- */
 export type Serializable = string;
 
 /** Represents a timestamp. */
@@ -31,28 +28,10 @@ class Timestamp {
     public toString(): string {
         return this.value.toString();
     }
-
-    /**
-     * Get a JSON-serializable representation of the timestamp.
-     * @returns {bigint} The JSON-serializable representation.
-     */
-    public toJSON(): bigint {
-        return this.value;
-    }
-}
-
-/**
- * Converts a `bigint` to timestamp.
- * @param {bigint} json The JSON representation of the timestamp.
- * @returns {Timestamp} The timestamp.
- */
-export function fromJSON(json: bigint): Timestamp {
-    return fromMillis(BigInt(json));
 }
 
 /**
  * Unwraps {@linkcode Type} value
- * @deprecated Use the {@linkcode Timestamp.toJSON} method instead.
  * @param value value to unwrap.
  * @returns the unwrapped {@linkcode bigint} value
  */
@@ -154,7 +133,6 @@ export function toProto(timestamp: Timestamp): Proto.Timestamp {
 
 /**
  * Constructs a {@linkcode Type} from {@linkcode Serializable}.
- * @deprecated Use the {@linkcode Timestamp.fromJSON} method instead.
  * @param {Serializable} value
  * @returns {Type} The duration.
  */
@@ -164,7 +142,6 @@ export function fromSerializable(value: Serializable): Type {
 
 /**
  * Converts {@linkcode Type} into {@linkcode Serializable}
- * @deprecated Use the {@linkcode Timestamp.toJSON} method instead.
  * @param {Type} value
  * @returns {Serializable} The serializable value
  */
@@ -174,7 +151,7 @@ export function toSerializable(value: Type): Serializable {
 
 /**
  * Takes an {@linkcode Type} and transforms it to a {@linkcode TypedJson} format.
- * @deprecated Use the {@linkcode Timestamp.toJSON} method instead.
+ * @deprecated Use the {@linkcode toSerializable} function instead.
  * @param {Type} value - The account address instance to transform.
  * @returns {TypedJson} The transformed object.
  */
@@ -187,7 +164,7 @@ export function toTypedJSON(value: Timestamp): TypedJson<Serializable> {
 
 /**
  * Takes a {@linkcode TypedJson} object and converts it to instance of type {@linkcode Type}.
- * @deprecated Use the{@linkcode fromJSON} function instead.
+ * @deprecated Use the{@linkcode fromSerializable} function instead.
  * @param {TypedJson} json - The typed JSON to convert.
  * @throws {TypedJsonParseError} - If unexpected JSON string is passed.
  * @returns {Type} The parsed instance.